requirements.txt
# Kisan Call Centre Query Assistant - Dependencies

# Web Framework
streamlit==1.31.0

# AI/ML Libraries
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.24.3

# IBM Watsonx Integration
ibm-watsonx-ai==0.2.6
ibm-cloud-sdk-core==3.18.2

# Data Processing
pandas==2.1.4

# Visualization
plotly==5.18.0

# Environment Management
python-dotenv==1.0.0

# Utilities
requests==2.31.0
